\name{EDCCGARCH-Data}
\alias{y_m5_2456}
\alias{y_m5_2581}
\alias{y_m17_3524}
\alias{theta_ini}
\title{Data set}
\description{
  y_m5_2456 is the dataset of the five major European stock market indices;
  y_m5_2581 is the dataset of the five major stock market indices;
  y_m17_3524 is the dataset of the 17 Fama-French industry portfolios. 
  theta_ini is a vector of initial values when the dimension m=5 and the order (r,s)=(1,0);
}
\author{Wenyu Li}

