\name{EDCCGARCH-package}
\alias{EDCCGARCH-package}
\alias{EDCCGARCH}
\docType{package}
\title{\packageTitle{EDCCGARCH}}
\description{\packageDescription{EDCCGARCH}}
\section{Package Content}{\packageIndices{EDCCGARCH}}
\author{\packageAuthor{EDCCGARCH}}
\section{Maintainer}{\packageMaintainer{EDCCGARCH}}
\keyword{package}
