% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCCtt.R
\name{DCCtt}
\alias{DCCtt}
\title{DCC-GARCH of Tse and Tsui (2002)}
\usage{
DCCtt(residuals, sigma, window.size = 1)
}
\arguments{
\item{residuals}{Residuals of the univariate GARCH models}

\item{sigma}{Conditional volatility of univariate GARCH models}

\item{m}{Window size. Default is 1.}
}
\value{
returns DCC-GARCH results
}
\description{
This function estimates the DCC-GARCH model of Tse and Tsui (2002)
}
\references{
Tse, Y. K., & Tsui, A. K. C. (2002). A multivariate generalized autoregressive conditional heteroscedasticity model with time-varying correlations. Journal of Business and Economic Statistics, 20(3), 351-362.
}
\author{
David Gabauer
}
